restructure-routes: put signin and signup page into (auth) and sharing same loading animation and layout
