+++ src/app/(auth)/layout.module.css
@@ -1,0 +1,7 @@
+.main {
+  height: 100%;
+  background-color: var(--color-blue-bg);
+  padding: 2.5rem 5rem;
+
+  position: relative;
+}
