+++ src/app/(auth)/loading.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import styles from "./loading.module.css";
+
+export default function Loading() {
+  return (
+    <div className={styles.spinnerContainer}>
+      <div className={styles.spinner}></div>
+    </div>
+  );
+}
