+++ src/app/(auth)/signin/page.tsx
@@ -1,15 +1,13 @@
-import React from "react";
-
+import React, { Suspense } from "react";
 import Image from "next/image";
-import Link from "next/link";
 import styles from "./page.module.css";
-import Navbar from "@/components/Navbar/Navbar";
 import Form from "@/components/Form/Form";
+import Link from "next/link";
+import Loading from "../loading";
 
-export default function Signup() {
+export default function Signin() {
   return (
-    <main className={styles.main}>
-      <Navbar />
+    <Suspense fallback={<Loading />}>
       <div className={styles.container}>
         <div className={styles.topbar}>
           <Image
@@ -19,15 +17,15 @@ export default function Signup() {
             height={60}
           />
           <div>
-            <h3 className="font-strong">Sign up</h3>
-            <p>Create an account to continue</p>
+            <h3 className="font-strong">Welcome back!</h3>
+            <p>Sign in to use your account</p>
           </div>
         </div>
-        <Form purpose="signup" />
+        <Form purpose="signin" />
         <div className={styles.guide_to_signup}>
-          <Link href="/signin">Already have a Readify account? Sign in</Link>
+          <Link href="/signup">Don&apos;t have an account? Sign up</Link>
         </div>
       </div>
-    </main>
+    </Suspense>
   );
 }
