+++ src/app/(auth)/signup/page.module.css
@@ -1,12 +1,4 @@
-.main {
-  height: 100%;
-  background-color: var(--color-blue-bg);
-  padding: 2.5rem 5rem;
-
-  position: relative;
-}
-
-.main .container {
+.container {
   position: absolute;
   top: 50%;
   left: 50%;
@@ -18,19 +10,19 @@
   padding: 3.5rem;
 }
 
-.main .container .topbar {
+.container .topbar {
   display: flex;
   align-items: center;
   gap: 1.8rem;
   margin-bottom: 5.2rem;
 }
 
-.main .container .guide_to_signup {
+.container .guide_to_signup {
   text-align: center;
   margin-top: 3.2rem;
   transition: all 0.5s;
 }
 
-.main .container .guide_to_signup:hover {
+.container .guide_to_signup:hover {
   text-decoration: underline;
 }
