+++ src/app/(auth)/signup/page.tsx
@@ -1,14 +1,14 @@
 import React, { Suspense } from "react";
-import Image from "next/image";
-import styles from "./page.module.css";
-import Navbar from "@/components/Navbar/Navbar";
-import Form from "@/components/Form/Form";
-import Link from "next/link";
 
-export default function Signin() {
+import Image from "next/image";
+import Link from "next/link";
+import styles from "./page.module.css";
+import Form from "@/components/Form/Form";
+import Loading from "../loading";
+
+export default function Signup() {
   return (
-    <main className={styles.main}>
-      <Navbar />
+    <Suspense fallback={<Loading />}>
       <div className={styles.container}>
         <div className={styles.topbar}>
           <Image
@@ -18,15 +18,15 @@ export default function Signin() {
             height={60}
           />
           <div>
-            <h3 className="font-strong">Welcome back!</h3>
-            <p>Sign in to use your account</p>
+            <h3 className="font-strong">Sign up</h3>
+            <p>Create an account to continue</p>
           </div>
         </div>
-        <Form purpose="signin" />
+        <Form purpose="signup" />
         <div className={styles.guide_to_signup}>
-          <Link href="/signup">Don&apos;t have an account? Sign up</Link>
+          <Link href="/signin">Already have a Readify account? Sign in</Link>
         </div>
       </div>
-    </main>
+    </Suspense>
   );
 }
